+++ src/utils.hpp
@@ -116,7 +116,7 @@ void saveRes(eval &evals, std::string time_dir)
     std::ofstream file(csvname);
     file << "idx,  img1_idx,  img2_idx,  num_pts,  time,  t_err,  R_err, T_GTnorm, rgt_lie\n";
     evals.average_time /= evals.time.size();
-    for (int i = 0; i < evals.R_err_per_round.size(); ++i)
+    for (size_t i = 0; i < evals.R_err_per_round.size(); ++i)
     {
         file << i + 1 << "," << evals.img_pair[i].first << "," << evals.img_pair[i].second << "," << evals.num_pts[i] << "," << evals.time[i] << "," << evals.t_err_per_round[i] << "," << evals.R_err_per_round[i] << "," << evals.t_gt_norm[i] << "," << evals.R_gt[i].transpose();
         // if (evals.noise[i] != 0)
